requirements.txt
fastapi~=0.110.3
uvicorn~=0.29.0
pydantic~=2.7.1
SQLAlchemy~=2.0.30
paho-mqtt~=2.1.0
python-dotenv~=1.0.1